Perf: change u64 to u32 for memory efficiency
Yeah, it saves 3.8MB for each 1,000,000 solves stored

+++ src/app.rs
@@ -441,12 +441,12 @@ impl cosmic::Application for AppModel {
 
             // TODO: refactor all this
             Message::TimerTick => {
-                self.timer.time = self.stopwatch.elapsed().as_millis() as u64;
+                self.timer.time = self.stopwatch.elapsed().as_millis() as u32;
             }
             Message::SpacePressed => {
                 self.space_pressed = true;
                 if self.timer.status == Status::Running {
-                    self.timer.time = self.stopwatch.elapsed().as_millis() as u64;
+                    self.timer.time = self.stopwatch.elapsed().as_millis() as u32;
                     let solve = Solve::new(self.timer.time, &self.current_scramble);
                     self.timer.status = Status::Stopped;
                     self.record.add_solve(solve);
